ValidParenthesis more optimized solution
